version changed open api added and other important parts

+++ spring-evento/src/main/java/com/example/spring_evento/service/EventoServiceImpl.java
@@ -38,7 +38,9 @@ public class EventoServiceImpl implements EventoService {
 
     @Override
     public void deleteEvento(Long id) {
-        eventoRepository.deleteById(id);
+        if (id != null) {
+            eventoRepository.deleteById(id);
+        }
     }
 
 }
